server/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
google==3.0.0
google-ai-generativelanguage==0.6.15
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
openai==2.28.0
packaging==26.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
PyMySQL==1.1.2
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.32.5
resend==2.25.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.6
